api/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9

# Environment
python-decouple>=3.8

# Production server
gunicorn>=22.0

Pillow>=10.0.0